Refs #29 (7) AttackCoordinator now uses calculateDistanceWeight(), and calculateStructureTypeWeight(). AttackCoordinator uses unscouted and unused expansions. Started creating a ResourceCounter, will mainly be used for AttackCoordinator to calculate how much resources an expansion has left (in calculateExpansionTimeWeight()).

+++ BATS/code/BATSModule/include/WaitGoal.h
@@ -63,8 +63,8 @@ protected:
 	GameTime* mpGameTime;
 
 private:
-	float mStartTime;
-	float mTimeout;
+	double mStartTime;
+	double mTimeout;
 	bool mUsesTimeout;
 };
 }
